services/cv-parser-service/requirements.txt
Flask
flask-cors
python-dotenv
werkzeug
PyPDF2
pdfplumber
openai
python-docx
PyJWT
requests
